+++ tunnel/tunnel.go
@@ -5,15 +5,15 @@ import (
 	"math/rand"
 	"time"
 
-	"github.com/go-kit/log"
+	"go.uber.org/zap"
 )
 
 type Tunnel struct {
-	log.Logger
+	*zap.Logger
 	Listener
 }
 
-func NewTunnel(logger log.Logger, l Listener) *Tunnel {
+func NewTunnel(logger *zap.Logger, l Listener) *Tunnel {
 	return &Tunnel{
 		Logger:   logger,
 		Listener: l,
@@ -23,10 +23,10 @@ func NewTunnel(logger log.Logger, l Listener) *Tunnel {
 func (t *Tunnel) Run(cnx *Connection) error {
 	ln, err := t.Listen()
 	if err != nil {
-		_ = t.Log("level", "fatal", "msg", "cannot listen on port", "err", err)
+		t.Error("cannot listen on port", zap.Error(err))
 		return err
 	}
-	_ = t.Log("msg", "Tunnel listening", "addr", t.Listener.Address())
+	t.Info("Tunnel listening", zap.String("addr", t.Listener.Address()))
 
 	defer ln.Close()
 
@@ -36,10 +36,10 @@ func (t *Tunnel) Run(cnx *Connection) error {
 		conn, err := ln.Accept()
 		if err != nil {
 			if err == io.EOF {
-				_ = t.Log("level", "warn", "msg", "connection closed")
+				t.Warn("connection closed")
 				break
 			}
-			_ = t.Log("level", "error", "msg", "error in listen", "err", err)
+			t.Error("error in listen", zap.Error(err))
 			continue
 		}
 		go cnx.handleConn(conn)